move creation of tweet link to a util
