+++ src/tweet-link/tweet-link.component.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import { bool, node, string } from 'prop-types';
 import { TwitterTweetEmbed } from 'react-twitter-embed';
+import ExternalLink from 'components/external-link';
 import './tweet-link.style.scss';
-
-// TODO: fix overlap with <LatestTweet/>
 
 export default class TweetLink extends React.Component {
   static propTypes = {
@@ -21,9 +20,9 @@ export default class TweetLink extends React.Component {
     }
 
     return (
-      <a className="tweetLink" href={`https://twitter.com/realDonaldTrump/status/${id}`} rel="noopener noreferrer" target="_blank">
+      <ExternalLink className="tweetLink" id={id}>
         { children || `"${text}"` }
-      </a>
+      </ExternalLink>
     );
   }
 }
