+++ src/utils/links.js
@@ -1,0 +1,1 @@
+export const tweetLink = id => `https://twitter.com/realdonaldtrump/status/${id}`;
